+++ custard/builder.py
@@ -559,4 +559,9 @@ class CustomFieldsBuilder(object):
             def __init__(self, *args, **kwargs):
                 super(CustomFieldModelBaseAdmin, self).__init__(*args, **kwargs)
 
+            def save_model(self, request, obj, form, change):
+                obj.save()
+                if hasattr(form, 'save_custom_fields'):
+                    form.save_custom_fields()
+
         return CustomFieldModelBaseAdmin